Failed attempt to merge training sets.
I created the file merge_and_simulate, but it is creating hdf5 files
that are too big. It might be a bug, but I am assuming that storing the
training dataset is not a good approach.

Now I am going to change training to have a seed and receive many
flights.  Simulations will still be created on demand.